fix: Move payload types to consistent location
