+++ app/src/types/payloads/PolicyCreationPayload.ts
@@ -1,0 +1,9 @@
+import { PolicyMetadataParams } from '@/types/metadata/policyMetadata';
+
+/**
+ * Payload format for creating a policy via the API
+ */
+export interface PolicyCreationPayload {
+  label?: string;
+  data: PolicyMetadataParams;
+}
