+++ app/src/types/payloads/ReportCreationPayload.ts
@@ -1,0 +1,7 @@
+/**
+ * Payload format for creating a report via the API
+ */
+export interface ReportCreationPayload {
+  simulation_id?: string;
+  // Add other fields as the API evolves
+}
