+++ app/src/types/payloads/SimulationCreationPayload.ts
@@ -1,0 +1,8 @@
+/**
+ * Payload format for creating a simulation via the API
+ */
+export interface SimulationCreationPayload {
+  population_id: string;
+  population_type?: 'household' | 'geography';
+  policy_id?: string;
+}
