--- app/src/types/simulationPayload.ts
@@ -1,15 +1,0 @@
-import { Simulation } from '@/types/ingredients/Simulation';
-
-export interface SimulationCreationPayload {
-  population_id?: string;
-  policy_id?: string;
-}
-
-export function serializeSimulationCreationPayload(
-  simulation: Simulation
-): SimulationCreationPayload {
-  return {
-    population_id: simulation.populationId,
-    policy_id: simulation.policyId,
-  };
-}
